+++ grpc.gyp
@@ -895,8 +895,8 @@
         'src/core/lib/debug/stats_data.cc',
         'src/core/lib/debug/trace.cc',
         'src/core/lib/event_engine/channel_args_endpoint_config.cc',
+        'src/core/lib/event_engine/default_event_engine_factory.cc',
         'src/core/lib/event_engine/event_engine.cc',
-        'src/core/lib/event_engine/event_engine_factory.cc',
         'src/core/lib/event_engine/memory_allocator.cc',
         'src/core/lib/event_engine/sockaddr.cc',
         'src/core/lib/http/format_request.cc',
@@ -1365,8 +1365,8 @@
         'src/core/lib/debug/stats_data.cc',
         'src/core/lib/debug/trace.cc',
         'src/core/lib/event_engine/channel_args_endpoint_config.cc',
+        'src/core/lib/event_engine/default_event_engine_factory.cc',
         'src/core/lib/event_engine/event_engine.cc',
-        'src/core/lib/event_engine/event_engine_factory.cc',
         'src/core/lib/event_engine/memory_allocator.cc',
         'src/core/lib/event_engine/sockaddr.cc',
         'src/core/lib/http/format_request.cc',
