+++ tools/doxygen/Doxyfile.c++.internal
@@ -1920,8 +1920,8 @@ src/core/lib/debug/trace.cc \
 src/core/lib/debug/trace.h \
 src/core/lib/event_engine/channel_args_endpoint_config.cc \
 src/core/lib/event_engine/channel_args_endpoint_config.h \
+src/core/lib/event_engine/default_event_engine_factory.cc \
 src/core/lib/event_engine/event_engine.cc \
-src/core/lib/event_engine/event_engine_factory.cc \
 src/core/lib/event_engine/event_engine_factory.h \
 src/core/lib/event_engine/memory_allocator.cc \
 src/core/lib/event_engine/sockaddr.cc \
